@@ -13,9 +13,9 @@
             "target-directory": "oxac/feefreepayments"
         }
     },
-    "version":  "1.0.0",
+    "version":  "1.0.1",
     "require": {
-        "php": ">=5.6"
+        "php": ">=7.1"
     },
     "autoload": {
         "psr-4": {
